minor changes making code modular
